+++ packages/amazonq/src/app/inline/EditRendering/displayImage.ts
@@ -12,6 +12,7 @@ import { LogInlineCompletionSessionResultsParams } from '@aws/language-server-ru
 import { InlineCompletionItemWithReferences } from '@aws/language-server-runtimes/protocol'
 import path from 'path'
 import { imageVerticalOffset } from './svgGenerator'
+import { AmazonQInlineCompletionItemProvider } from '../completion'
 
 export class EditDecorationManager {
     private imageDecorationType: vscode.TextEditorDecorationType
@@ -280,8 +281,7 @@ export async function displaySvgDecoration(
     session: CodeWhispererSession,
     languageClient: LanguageClient,
     item: InlineCompletionItemWithReferences,
-    addedCharacterCount: number,
-    deletedCharacterCount: number
+    inlineCompletionProvider?: AmazonQInlineCompletionItemProvider
 ) {
     const originalCode = editor.document.getText()
 
@@ -289,7 +289,7 @@ export async function displaySvgDecoration(
         editor,
         svgImage,
         startLine,
-        () => {
+        async () => {
             // Handle accept
             getLogger().info('Edit suggestion accepted')
 
@@ -315,10 +315,20 @@ export async function displaySvgDecoration(
                 },
                 totalSessionDisplayTime: Date.now() - session.requestStartTime,
                 firstCompletionDisplayLatency: session.firstCompletionDisplayLatency,
-                addedCharacterCount: addedCharacterCount,
-                deletedCharacterCount: deletedCharacterCount,
             }
             languageClient.sendNotification('aws/logInlineCompletionSessionResults', params)
+            if (inlineCompletionProvider) {
+                await inlineCompletionProvider.provideInlineCompletionItems(
+                    editor.document,
+                    endPosition,
+                    {
+                        triggerKind: vscode.InlineCompletionTriggerKind.Automatic,
+                        selectedCompletionInfo: undefined,
+                    },
+                    new vscode.CancellationTokenSource().token,
+                    { emitTelemetry: false, showUi: false }
+                )
+            }
         },
         () => {
             // Handle reject
@@ -333,8 +343,6 @@ export async function displaySvgDecoration(
                         discarded: false,
                     },
                 },
-                // addedCharacterCount: addedCharacterCount,
-                // deletedCharacterCount: deletedCharacterCount,
             }
             languageClient.sendNotification('aws/logInlineCompletionSessionResults', params)
         },